keiba-ai-ver1/pyproject.toml
[project]
name = "keiba-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Data collection
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "playwright>=1.40.0",
    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0",
    # Machine learning
    "lightgbm>=4.1.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # Configuration
    "pyyaml>=6.0.1",
    # Utilities
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
]
